requirements.txt
python-telegram-bot>=22.5
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
aiofiles>=23.2.1
yt-dlp>=2025.1.0
